chore: board 관련 dto, model 클래스를 하위 board 패키지로 이동
