+++ CatchMate/data/src/main/java/com/catchmate/data/datasource/local/GoogleLoginDataSource.kt
@@ -9,7 +9,7 @@ import androidx.credentials.GetCredentialResponse
 import com.catchmate.data.BuildConfig
 import com.catchmate.data.datasource.remote.FCMTokenService
 import com.catchmate.data.dto.PostLoginRequestDTO
-import com.catchmate.domain.model.LoginPlatform
+import com.catchmate.domain.model.enumclass.LoginPlatform
 import com.google.android.libraries.identity.googleid.GetGoogleIdOption
 import com.google.android.libraries.identity.googleid.GoogleIdTokenCredential
 import com.google.android.libraries.identity.googleid.GoogleIdTokenParsingException
