+++ CatchMate/data/src/main/java/com/catchmate/data/datasource/local/NaverLoginDataSource.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.util.Log
 import com.catchmate.data.datasource.remote.FCMTokenService
 import com.catchmate.data.dto.PostLoginRequestDTO
-import com.catchmate.domain.model.LoginPlatform
+import com.catchmate.domain.model.enumclass.LoginPlatform
 import com.navercorp.nid.NaverIdLoginSDK
 import com.navercorp.nid.oauth.NidOAuthLogin
 import com.navercorp.nid.oauth.OAuthLoginCallback
