--- CatchMate/data/src/main/java/com/catchmate/data/dto/GetBoardPagingResponseDTO.kt
@@ -1,13 +1,0 @@
-package com.catchmate.data.dto
-
-data class GetBoardPagingResponseDTO(
-    val boardId: Long,
-    val title: String,
-    val gameDate: String,
-    val location: String,
-    val homeTeam: String,
-    val awayTeam: String,
-    val cheerTeam: String,
-    val currentPerson: Int,
-    val maxPerson: Int,
-)
