+++ CatchMate/data/src/main/java/com/catchmate/data/dto/board/DeleteBoardRequestDTO.kt
@@ -1,4 +1,4 @@
-package com.catchmate.data.dto
+package com.catchmate.data.dto.board
 
 data class DeleteBoardRequestDTO(
     val boardId: Long,
