+++ CatchMate/data/src/main/java/com/catchmate/data/dto/board/GetBoardPagingResponseDTO.kt
@@ -1,6 +1,6 @@
-package com.catchmate.domain.model
+package com.catchmate.data.dto.board
 
-data class GetBoardPagingResponse(
+data class GetBoardPagingResponseDTO(
     val boardId: Long,
     val title: String,
     val gameDate: String,
