+++ CatchMate/data/src/main/java/com/catchmate/data/dto/board/GetBoardResponseDTO.kt
@@ -1,13 +1,16 @@
-package com.catchmate.domain.model
+package com.catchmate.data.dto.board
 
-data class PostBoardRequest(
+data class GetBoardResponseDTO(
+    val boardId: Long,
+    val writer: WriterDTO,
     val title: String,
     val gameDate: String,
     val location: String,
     val homeTeam: String,
     val awayTeam: String,
-    val maxPerson: Int,
     val cheerTeam: String,
+    val maxPerson: Int,
+    val currentPerson: Int,
     val preferGender: String? = null,
     val preferAge: Int? = null,
     val addInfo: String,
