+++ CatchMate/data/src/main/java/com/catchmate/data/dto/board/PostBoardRequestDTO.kt
@@ -1,0 +1,14 @@
+package com.catchmate.data.dto.board
+
+data class PostBoardRequestDTO(
+    val title: String,
+    val gameDate: String,
+    val location: String,
+    val homeTeam: String,
+    val awayTeam: String,
+    val maxPerson: Int,
+    val cheerTeam: String,
+    val preferGender: String? = null,
+    val preferAge: Int? = null,
+    val addInfo: String,
+)
