+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/board/GetLikedBoardResponse.kt
@@ -1,6 +1,6 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.board
 
-data class PutBoardRequest(
+data class GetLikedBoardResponse(
     val boardId: Long,
     val title: String,
     val gameDate: String,
@@ -10,7 +10,4 @@ data class PutBoardRequest(
     val cheerTeam: String,
     val currentPerson: Int,
     val maxPerson: Int,
-    val preferGender: String? = null,
-    val preferAge: Int? = null,
-    val addInfo: String,
 )
