+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/board/PostBoardRequest.kt
@@ -1,19 +1,14 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.board
 
-import java.io.Serializable
-
-data class GetBoardResponse(
-    val boardId: Long,
-    val writer: Writer,
+data class PostBoardRequest(
     val title: String,
     val gameDate: String,
     val location: String,
     val homeTeam: String,
     val awayTeam: String,
-    val cheerTeam: String,
     val maxPerson: Int,
-    val currentPerson: Int,
+    val cheerTeam: String,
     val preferGender: String? = null,
     val preferAge: Int? = null,
     val addInfo: String,
-) : Serializable
+)
