+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/board/PostBoardResponse.kt
@@ -1,4 +1,4 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.board
 
 data class PostBoardResponse(
     val boardId: Long,
