+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/board/PutBoardRequest.kt
@@ -1,13 +1,15 @@
-package com.catchmate.data.dto
+package com.catchmate.domain.model.board
 
-data class PostBoardRequestDTO(
+data class PutBoardRequest(
+    val boardId: Long,
     val title: String,
     val gameDate: String,
     val location: String,
     val homeTeam: String,
     val awayTeam: String,
-    val maxPerson: Int,
     val cheerTeam: String,
+    val currentPerson: Int,
+    val maxPerson: Int,
     val preferGender: String? = null,
     val preferAge: Int? = null,
     val addInfo: String,
