+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/board/PutBoardResponse.kt
@@ -1,4 +1,4 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.board
 
 data class PutBoardResponse(
     val boardId: Long,
