+++ CatchMate/domain/src/main/java/com/catchmate/domain/model/enumclass/EnrollState.kt
@@ -1,4 +1,4 @@
-package com.catchmate.domain.model
+package com.catchmate.domain.model.enumclass
 
 enum class EnrollState {
     APPLICABLE, // 신청 가능
