+++ CatchMate/domain/src/main/java/com/catchmate/domain/usecase/board/PutBoardUseCase.kt
@@ -1,7 +1,7 @@
 package com.catchmate.domain.usecase.board
 
-import com.catchmate.domain.model.PutBoardRequest
-import com.catchmate.domain.model.PutBoardResponse
+import com.catchmate.domain.model.board.PutBoardRequest
+import com.catchmate.domain.model.board.PutBoardResponse
 import com.catchmate.domain.repository.BoardRepository
 import javax.inject.Inject
